pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "nftpol"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[project.optional-dependencies]
docker = ["docker"]
dev = ["pytest"]

[project.scripts]
nftpol = "nftpol.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nftpol*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
